# anms-core/integration_test/openapi.json
{
  "openapi": "3.0.2",
  "info": {
      "title": "AMMOS-ANMS",
      "description": "AMMOS-ANMS Docs",
      "version": "0.1.0"
  },
  "paths": {
...
      "/ari": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Ari",
              "operationId": "paged_ARI_ari_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_ARI_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/ari/name/{obj_name}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Ari By Name",
              "operationId": "ari_by_name_ari_name__obj_name__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Name",
                          "type": "string"
                      },
                      "name": "obj_name",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ARI"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/actual_objects": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Actual Object",
              "operationId": "paged_actual_object_actual_objects_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_ActualObject_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/actual_objects/id/{obj_metadata_id}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Actual Object By Id",
              "operationId": "actual_object_by_id_actual_objects_id__obj_metadata_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Metadata Id",
                          "type": "integer"
                      },
                      "name": "obj_metadata_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ActualObject"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
      "/actual_objects/name/{obj_name}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Actual Object By Name",
              "operationId": "actual_object_by_name_actual_objects_name__obj_name__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Obj Name",
                          "type": "string"
                      },
                      "name": "obj_name",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ActualObject"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/actual_parameter": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Paged Actualparameter",
              "operationId": "paged_ActualParameter_actual_parameter_get",
              "parameters": [
                  {
                      "required": false,
                      "schema": {
                          "title": "Page",
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 1
                      },
                      "name": "page",
                      "in": "query"
                  },
                  {
                      "required": false,
                      "schema": {
                          "title": "Size",
                          "maximum": 100.0,
                          "minimum": 1.0,
                          "type": "integer",
                          "default": 50
                      },
                      "name": "size",
                      "in": "query"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Page_ActualParameter_"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
      "/actual_parameter/id/{ap_spec_id}": {
          "get": {
              "tags": [
                  "ARIs"
              ],
              "summary": "Actual Parameter By Id",
              "operationId": "actual_parameter_by_id_actual_parameter_id__ap_spec_id__get",
              "parameters": [
                  {
                      "required": true,
                      "schema": {
                          "title": "Ap Spec Id",
                          "type": "integer"
                      },
                      "name": "ap_spec_id",
                      "in": "path"
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful Response",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ActualParameter"
                              }
                          }
                      }
                  },
                  "422": {
                      "description": "Validation Error",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/HTTPValidationError"
                              }
                          }
                      }
                  }
              }
          }
      },
...
  },
  "components": {
      "schemas": {
          "ARI": {
              "title": "ARI",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "type_name": {
                      "title": "Type Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "obj_id": {
                      "title": "Obj Id",
                      "type": "integer"
                  },
                  "parm_id": {
                      "title": "Parm Id",
                      "type": "integer"
                  },
                  "actual": {
                      "title": "Actual",
                      "type": "boolean"
                  },
                  "display": {
                      "title": "Display",
                      "type": "string"
                  }
              }
          },
...
          "ActualObject": {
              "title": "ActualObject",
              "type": "object",
              "properties": {
                  "obj_metadata_id": {
                      "title": "Obj Metadata Id",
                      "type": "integer"
                  },
                  "type_name": {
                      "title": "Type Name",
                      "type": "string"
                  },
                  "obj_name": {
                      "title": "Obj Name",
                      "type": "string"
                  },
                  "namespace_id": {
                      "title": "Namespace Id",
                      "type": "integer"
                  },
                  "adm_name": {
                      "title": "Adm Name",
                      "type": "string"
                  },
                  "adm_enum": {
                      "title": "Adm Enum",
                      "type": "integer"
                  },
                  "adm_enum_label": {
                      "title": "Adm Enum Label",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  },
                  "obj_actual_definition_id": {
                      "title": "Obj Actual Definition Id",
                      "type": "integer"
                  },
                  "actual_desc": {
                      "title": "Actual Desc",
                      "type": "string"
                  },
                  "is_actual": {
                      "title": "Is Actual",
                      "type": "boolean",
                      "default": true
                  }
              }
          },
          "ActualParameter": {
              "title": "ActualParameter",
              "type": "object",
              "properties": {
                  "ap_spec_id": {
                      "title": "Ap Spec Id",
                      "type": "integer"
                  },
                  "tnvc_id": {
                      "title": "Tnvc Id",
                      "type": "integer"
                  },
                  "fp_spec_id": {
                      "title": "Fp Spec Id",
                      "type": "integer"
                  },
                  "use_desc": {
                      "title": "Use Desc",
                      "type": "string"
                  },
                  "fp_values": {
                      "title": "Fp Values",
                      "type": "string"
                  },
                  "str_values": {
                      "title": "Str Values",
                      "type": "string"
                  },
                  "uint_values": {
                      "title": "Uint Values",
                      "type": "string"
                  },
                  "int_values": {
                      "title": "Int Values",
                      "type": "string"
                  },
                  "obj_values": {
                      "title": "Obj Values",
                      "type": "string"
                  },
                  "ac_values": {
                      "title": "Ac Values",
                      "type": "string"
                  },
                  "tnvc_values": {
                      "title": "Tnvc Values",
                      "type": "string"
                  },
                  "real32_values": {
                      "title": "Real32 Values",
                      "type": "string"
                  },
                  "real64_values": {
                      "title": "Real64 Values",
                      "type": "string"
                  },
                  "uvast_values": {
                      "title": "Uvast Values",
                      "type": "string"
                  },
                  "vast_values": {
                      "title": "Vast Values",
                      "type": "string"
                  }
              }
          },
...
          "HTTPValidationError": {
              "title": "HTTPValidationError",
              "type": "object",
              "properties": {
                  "detail": {
                      "title": "Detail",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ValidationError"
                      }
                  }
              }
          },
...
          "Page_ARI_": {
              "title": "Page[ARI]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ARI"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
          "Page_ActualObject_": {
              "title": "Page[ActualObject]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ActualObject"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
          "Page_ActualParameter_": {
              "title": "Page[ActualParameter]",
              "required": [
                  "items",
                  "total",
                  "page",
                  "size"
              ],
              "type": "object",
              "properties": {
                  "items": {
                      "title": "Items",
                      "type": "array",
                      "items": {
                          "$ref": "#/components/schemas/ActualParameter"
                      }
                  },
                  "total": {
                      "title": "Total",
                      "minimum": 0.0,
                      "type": "integer"
                  },
                  "page": {
                      "title": "Page",
                      "minimum": 1.0,
                      "type": "integer"
                  },
                  "size": {
                      "title": "Size",
                      "minimum": 1.0,
                      "type": "integer"
                  }
              }
          },
...
          "ValidationError": {
              "title": "ValidationError",
              "required": [
                  "loc",
                  "msg",
                  "type"
              ],
              "type": "object",
              "properties": {
                  "loc": {
                      "title": "Location",
                      "type": "array",
                      "items": {
                          "anyOf": [
                              {
                                  "type": "string"
                              },
                              {
                                  "type": "integer"
                              }
                          ]
                      }
                  },
                  "msg": {
                      "title": "Message",
                      "type": "string"
                  },
                  "type": {
                      "title": "Error Type",
                      "type": "string"
                  }
              }
          },
...
      }
  }
}